+++ CHANGELOG.md
@@ -8,9 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+=======
 ## Added
-
+- Can use both the `state` and `description` property for the `Autocomplete` component
 - Can use state and description in the `inputProps` for the Datepicker component
 
 ## [7.0.1] - 2023-08-10